refactor(lsposed): split DiagnosticsScreen into UI, checks, and export
DiagnosticsScreen.kt mixed three concerns in ~1240 lines. Extract the
~20 Java/native probe functions into JavaChecks.kt and the debug-zip /
device-info gathering into DebugExport.kt, leaving DiagnosticsScreen.kt
as the screen UI. No behaviour change; exportDebugZip becomes internal
so the UI section can still call it across the file boundary.
